+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,7 +1,11 @@
-
 ## 2.2.0
 
 * Adds options for gesture handling and tilt controls on web.
+
+## 2.1.6
+
+* Migrates from `ui.hash*` to `Object.hash*`.
+* Updates minimum Flutter version to 2.5.0.
 
 ## 2.1.5